requirements.txt
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
pandas @ file:///C:/ci/pandas_1618365634936/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
streamlit==1.2.0
Pillow @ file:///C:/ci/pillow_1617386341487/work